Move MongoDB driver to new namespace
